requirements.txt
absl-py==1.0.0
anyio==3.3.4
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
boto3==1.34.117
botocore==1.34.117
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
click==8.1.7
conda==4.10.3
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.14.0
fsspec==2024.5.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.6
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyter-server==1.11.2
jupyterlab==3.2.3
jupyterlab-language-pack-zh-CN==3.2.post0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.3.0
nbclassic==0.3.4
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==3.1
notebook==6.4.5
numpy==1.19.5
nvidia-cublas==11.5.1.101
nvidia-cublas-cu114==11.6.5.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu114==11.4.120
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu114==11.4.120
nvidia-cuda-nvrtc==11.3.58
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==11.3.58
nvidia-cuda-runtime-cu114==11.4.108
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu114==8.2.4.15
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu114==10.5.2.100
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu114==10.2.5.120
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu114==11.2.0.120
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu114==11.6.0.120
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda110==1.6.0
nvidia-dali-nvtf-plugin==1.6.0+nv21.10
nvidia-nccl-cu114==2.11.4
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
nvidia-pyindex==1.0.9
nvidia-tensorflow==1.15.5+nv21.10
nvidia-tensorrt==8.0.3.4
opt-einsum==3.3.0
packaging==21.2
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.10.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
regex==2024.5.15
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
s3transfer==0.10.1
sacremoses==0.1.1
Send2Trash==1.8.0
sentencepiece==0.2.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sniffio==1.2.0
supervisor==4.2.2
sympy==1.12.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tokenizers==0.5.2
torch==2.3.0
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.1.1
transformers==2.5.1
triton==2.3.0
typing-extensions==4.12.1
tzdata==2024.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.6.0
